+++ neopdf/src/interpolator.rs
@@ -156,6 +156,7 @@ impl InterpolatorFactory {
         pid_index: usize,
     ) -> Box<dyn DynInterpolator> {
         let grid_slice = subgrid.grid_slice(pid_index).to_owned();
+
         match interp_type {
             InterpolatorType::Bilinear => Box::new(
                 Interp2D::new(
@@ -169,8 +170,8 @@ impl InterpolatorFactory {
             ),
             InterpolatorType::LogBilinear => Box::new(
                 Interp2D::new(
-                    subgrid.xs.to_owned(),
-                    subgrid.q2s.to_owned(),
+                    subgrid.xs.mapv(f64::ln),
+                    subgrid.q2s.mapv(f64::ln),
                     grid_slice,
                     LogBilinearInterpolation,
                     Extrapolate::Clamp,
@@ -179,8 +180,8 @@ impl InterpolatorFactory {
             ),
             InterpolatorType::LogBicubic => Box::new(
                 Interp2D::new(
-                    subgrid.xs.to_owned(),
-                    subgrid.q2s.to_owned(),
+                    subgrid.xs.mapv(f64::ln),
+                    subgrid.q2s.mapv(f64::ln),
                     grid_slice,
                     LogBicubicInterpolation::default(),
                     Extrapolate::Clamp,
@@ -189,8 +190,8 @@ impl InterpolatorFactory {
             ),
             InterpolatorType::LogChebyshev => Box::new(
                 Interp2D::new(
-                    subgrid.xs.to_owned(),
-                    subgrid.q2s.to_owned(),
+                    subgrid.xs.mapv(f64::ln),
+                    subgrid.q2s.mapv(f64::ln),
                     grid_slice,
                     LogChebyshevInterpolation::<2>::default(),
                     Extrapolate::Clamp,
@@ -213,12 +214,13 @@ impl InterpolatorFactory {
         let reshaped_data = grid_data
             .into_shape_with_order((subgrid.nucleons.len(), subgrid.xs.len(), subgrid.q2s.len()))
             .expect("Failed to reshape 3D data");
+
         match interp_type {
             InterpolatorType::LogTricubic => Box::new(
                 Interp3D::new(
-                    subgrid.nucleons.to_owned(),
-                    subgrid.xs.to_owned(),
-                    subgrid.q2s.to_owned(),
+                    subgrid.nucleons.mapv(f64::ln),
+                    subgrid.xs.mapv(f64::ln),
+                    subgrid.q2s.mapv(f64::ln),
                     reshaped_data,
                     LogTricubicInterpolation,
                     Extrapolate::Clamp,
@@ -227,9 +229,9 @@ impl InterpolatorFactory {
             ),
             InterpolatorType::LogChebyshev => Box::new(
                 Interp3D::new(
-                    subgrid.nucleons.to_owned(),
-                    subgrid.xs.to_owned(),
-                    subgrid.q2s.to_owned(),
+                    subgrid.nucleons.mapv(f64::ln),
+                    subgrid.xs.mapv(f64::ln),
+                    subgrid.q2s.mapv(f64::ln),
                     reshaped_data,
                     LogChebyshevInterpolation::<3>::default(),
                     Extrapolate::Clamp,
@@ -252,12 +254,13 @@ impl InterpolatorFactory {
         let reshaped_data = grid_data
             .into_shape_with_order((subgrid.alphas.len(), subgrid.xs.len(), subgrid.q2s.len()))
             .expect("Failed to reshape 3D data");
+
         match interp_type {
             InterpolatorType::LogTricubic => Box::new(
                 Interp3D::new(
-                    subgrid.alphas.to_owned(),
-                    subgrid.xs.to_owned(),
-                    subgrid.q2s.to_owned(),
+                    subgrid.alphas.mapv(f64::ln),
+                    subgrid.xs.mapv(f64::ln),
+                    subgrid.q2s.mapv(f64::ln),
                     reshaped_data,
                     LogTricubicInterpolation,
                     Extrapolate::Clamp,
@@ -266,9 +269,9 @@ impl InterpolatorFactory {
             ),
             InterpolatorType::LogChebyshev => Box::new(
                 Interp3D::new(
-                    subgrid.alphas.to_owned(),
-                    subgrid.xs.to_owned(),
-                    subgrid.q2s.to_owned(),
+                    subgrid.alphas.mapv(f64::ln),
+                    subgrid.xs.mapv(f64::ln),
+                    subgrid.q2s.mapv(f64::ln),
                     reshaped_data,
                     LogChebyshevInterpolation::<3>::default(),
                     Extrapolate::Clamp,
@@ -291,12 +294,13 @@ impl InterpolatorFactory {
         let reshaped_data = grid_data
             .into_shape_with_order((subgrid.kts.len(), subgrid.xs.len(), subgrid.q2s.len()))
             .expect("Failed to reshape 3D data");
+
         match interp_type {
             InterpolatorType::LogTricubic => Box::new(
                 Interp3D::new(
-                    subgrid.kts.to_owned(),
-                    subgrid.xs.to_owned(),
-                    subgrid.q2s.to_owned(),
+                    subgrid.kts.mapv(f64::ln),
+                    subgrid.xs.mapv(f64::ln),
+                    subgrid.q2s.mapv(f64::ln),
                     reshaped_data,
                     LogTricubicInterpolation,
                     Extrapolate::Clamp,
@@ -305,9 +309,9 @@ impl InterpolatorFactory {
             ),
             InterpolatorType::LogChebyshev => Box::new(
                 Interp3D::new(
-                    subgrid.kts.to_owned(),
-                    subgrid.xs.to_owned(),
-                    subgrid.q2s.to_owned(),
+                    subgrid.kts.mapv(f64::ln),
+                    subgrid.xs.mapv(f64::ln),
+                    subgrid.q2s.mapv(f64::ln),
                     reshaped_data,
                     LogChebyshevInterpolation::<3>::default(),
                     Extrapolate::Clamp,
@@ -341,6 +345,7 @@ impl InterpolatorFactory {
                 subgrid.q2s.len(),
             ))
             .expect("Failed to reshape 4D data");
+
         match interp_type {
             InterpolatorType::InterpNDLinear => Box::new(
                 InterpND::new(coords, reshaped_data.into_dyn(), Linear, Extrapolate::Clamp)
@@ -360,10 +365,10 @@ impl InterpolatorFactory {
             .slice(s![.., 0, pid_index, .., .., ..])
             .to_owned();
         let coords = vec![
-            subgrid.nucleons.to_owned(),
-            subgrid.kts.to_owned(),
-            subgrid.xs.to_owned(),
-            subgrid.q2s.to_owned(),
+            subgrid.nucleons.mapv(f64::ln),
+            subgrid.kts.mapv(f64::ln),
+            subgrid.xs.mapv(f64::ln),
+            subgrid.q2s.mapv(f64::ln),
         ];
         let reshaped_data = grid_data
             .into_shape_with_order((
@@ -373,6 +378,7 @@ impl InterpolatorFactory {
                 subgrid.q2s.len(),
             ))
             .expect("Failed to reshape 4D data");
+
         match interp_type {
             InterpolatorType::InterpNDLinear => Box::new(
                 InterpND::new(coords, reshaped_data.into_dyn(), Linear, Extrapolate::Clamp)
@@ -392,10 +398,10 @@ impl InterpolatorFactory {
             .slice(s![0, .., pid_index, .., .., ..])
             .to_owned();
         let coords = vec![
-            subgrid.alphas.to_owned(),
-            subgrid.kts.to_owned(),
-            subgrid.xs.to_owned(),
-            subgrid.q2s.to_owned(),
+            subgrid.alphas.mapv(f64::ln),
+            subgrid.kts.mapv(f64::ln),
+            subgrid.xs.mapv(f64::ln),
+            subgrid.q2s.mapv(f64::ln),
         ];
         let reshaped_data = grid_data
             .into_shape_with_order((
@@ -405,6 +411,7 @@ impl InterpolatorFactory {
                 subgrid.q2s.len(),
             ))
             .expect("Failed to reshape 4D data");
+
         match interp_type {
             InterpolatorType::InterpNDLinear => Box::new(
                 InterpND::new(coords, reshaped_data.into_dyn(), Linear, Extrapolate::Clamp)
@@ -424,11 +431,11 @@ impl InterpolatorFactory {
             .slice(s![.., .., pid_index, .., .., ..])
             .to_owned();
         let coords = vec![
-            subgrid.nucleons.to_owned(),
-            subgrid.alphas.to_owned(),
-            subgrid.kts.to_owned(),
-            subgrid.xs.to_owned(),
-            subgrid.q2s.to_owned(),
+            subgrid.nucleons.mapv(f64::ln),
+            subgrid.alphas.mapv(f64::ln),
+            subgrid.kts.mapv(f64::ln),
+            subgrid.xs.mapv(f64::ln),
+            subgrid.q2s.mapv(f64::ln),
         ];
         let reshaped_data = grid_data
             .into_shape_with_order((
@@ -439,6 +446,7 @@ impl InterpolatorFactory {
                 subgrid.q2s.len(),
             ))
             .expect("Failed to reshape 5D data");
+
         match interp_type {
             InterpolatorType::InterpNDLinear => Box::new(
                 InterpND::new(coords, reshaped_data.into_dyn(), Linear, Extrapolate::Clamp)
@@ -454,7 +462,7 @@ mod tests {
     use super::*;
     use crate::subgrid::SubGrid;
 
-    const MAXDIFF: f64 = 1e-16;
+    const MAXDIFF: f64 = 1e-15;
 
     fn mock_subgrid_2d() -> SubGrid {
         let xs = vec![0.1, 0.2];
@@ -546,7 +554,9 @@ mod tests {
     fn test_3d_nucleons_interpolation() {
         let subgrid = mock_subgrid_3d_nucleons();
         let interpolator = InterpolatorFactory::create(InterpolatorType::LogTricubic, &subgrid, 0);
-        let result = interpolator.interpolate_point(&[2.0, 0.2, 2.0]).unwrap();
+        let result = interpolator
+            .interpolate_point(&[2.0f64.ln(), 0.2f64.ln(), 2.0f64.ln()])
+            .unwrap();
         assert!((result - 22.0).abs() < MAXDIFF);
     }
 
@@ -554,7 +564,9 @@ mod tests {
     fn test_3d_alphas_interpolation() {
         let subgrid = mock_subgrid_3d_alphas();
         let interpolator = InterpolatorFactory::create(InterpolatorType::LogTricubic, &subgrid, 0);
-        let result = interpolator.interpolate_point(&[0.120, 0.2, 2.0]).unwrap();
+        let result = interpolator
+            .interpolate_point(&[0.120f64.ln(), 0.2f64.ln(), 2.0f64.ln()])
+            .unwrap();
         assert!((result - 22.0).abs() < MAXDIFF);
     }
 
@@ -562,7 +574,9 @@ mod tests {
     fn test_3d_kts_interpolation() {
         let subgrid = mock_subgrid_3d_kts();
         let interpolator = InterpolatorFactory::create(InterpolatorType::LogTricubic, &subgrid, 0);
-        let result = interpolator.interpolate_point(&[1.0, 0.2, 2.0]).unwrap();
+        let result = interpolator
+            .interpolate_point(&[1.0f64.ln(), 0.2f64.ln(), 2.0f64.ln()])
+            .unwrap();
         assert!((result - 22.0).abs() < MAXDIFF);
     }